Move folders correctly
