+++ src/components/PetRooms/ShowRoom.js
@@ -2,8 +2,8 @@ import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useParams } from 'react-router-dom';
 import { fetchRoomId } from '../../redux/thunks/fetchRooms';
-import CalculateRating from '../Extra/CalculateRating';
-import HomeCarousel from '../Extra/Carousel';
+import CalculateRating from '../../utils/CalculateRating';
+import HomeCarousel from './Carousel';
 import ModalRoot from '../Modal/ModalRoot';
 import ModalService from '../Modal/ModalService';
 import MakeReservation from '../Reservations/MakeReservation';
